522: v1.13: add get_last_check_results
